Reduce coalescing delay for bulk data transfers
On low-traffic connections, blocks can be delayed by up to 5 seconds, and
it appears this can accumulate on further low-traffic connections. Hence
in my simulation most inserts take ~ 30 seconds, but some time >5 minutes.
I think this is why inserts are so slow on the main network: They have to
wait for the slowest downstream transfer, and if it's a connection to a
newbie node, it might have no other traffic.